'#1058 [WIP] Config section. Component yaml'
