+++ package.json
@@ -130,7 +130,8 @@
     "webext-dynamic-content-scripts": "^8.0.0",
     "webext-patterns": "^1.1.1",
     "webext-polyfill-kinda": "^0.1.0",
-    "webextension-polyfill-ts": "^0.26.0"
+    "webextension-polyfill-ts": "^0.26.0",
+    "yaml": "^1.10.2"
   },
   "devDependencies": {
     "@babel/core": "^7.14.6",
