+++ src/options/pages/brickEditor/referenceEntryType.tsx
@@ -1,0 +1,3 @@
+import { IBlock, IExtensionPoint, IService } from "@/core";
+
+export type ReferenceEntry = IBlock | IExtensionPoint | IService;
